changes provided in EESSI PR 496
