+++ create_tarball.sh
@@ -51,6 +51,11 @@ if [ -d ${eessi_version}/scripts ]; then
     find ${eessi_version}/scripts -type f | grep -v '/\.wh\.' >> ${files_list}
 fi
 
+# also include init, which is also copied by install_scripts.sh
+if [ -d ${eessi_version}/init ]; then
+    find ${eessi_version}/init -type f | grep -v '/\.wh\.' >> ${files_list}
+fi
+
 if [ -d ${eessi_version}/software/${os}/${cpu_arch_subdir}/modules ]; then
     # module files
     find ${eessi_version}/software/${os}/${cpu_arch_subdir}/modules -type f | grep -v '/\.wh\.' >> ${files_list}
